+++ libs/cmcd/src/fromCmcdHeaders.ts
@@ -1,8 +1,7 @@
 import type { Cmcd } from './Cmcd.ts'
-import { CMCD_OBJECT, CMCD_REQUEST, CMCD_SESSION, CMCD_STATUS } from './CmcdHeaderField.ts'
+import { CMCD_HEADER_FIELDS } from './CmcdHeaderField.ts'
 import { decodeCmcd } from './decodeCmcd.ts'
 
-const keys = [CMCD_OBJECT, CMCD_REQUEST, CMCD_SESSION, CMCD_STATUS]
 
 /**
  * Decode CMCD data from request headers.
@@ -21,7 +20,7 @@ export function fromCmcdHeaders(headers: Record<string, string> | Headers): Cmcd
 		headers = new Headers(headers)
 	}
 
-	return keys.reduce((acc, key) => {
+	return CMCD_HEADER_FIELDS.reduce((acc, key) => {
 		const value = headers.get(key)
 		return Object.assign(acc, decodeCmcd(value as string))
 	}, {} as Cmcd)